delete now checks if you really want to delete the key

+++ cmd/delete.go
@@ -1,8 +1,10 @@
 package cmd
 
 import (
+	"bufio"
 	"fmt"
 	"strings"
+	"os"
 
 	"github.com/Alvesafk/gopass/color"
 	"github.com/Alvesafk/gopass/storage"
@@ -32,13 +34,32 @@ func Delete(db storage.DB, args []string) {
 		return
 	}
 
-	err = db.Delete(to_get_secret.ID)
+	reader := bufio.NewReader(os.Stdin)
+
+	fmt.Printf("Are you sure you want to delete %s? y/N : ", to_get_secret.Name)
+	response, err := reader.ReadString('\n')
 	if err != nil {
-		fmt.Printf(color.Red("Error: Could not delete %s.", "bold", 1), to_get_name)
 		fmt.Println(err)
-		return
 	}
 
-	fmt.Printf(color.Green("Success! %s secret was deleted.", "bold", 1), args[2])
+	switch strings.TrimSpace(response) {
+	case "Y", "y", "yes", "Yes", "YES":
+		err = db.Delete(to_get_secret.ID)
+		if err != nil {
+			fmt.Printf(color.Red("Error: Could not delete %s.", "bold", 1), to_get_name)
+			fmt.Println(err)
+			return
+		}
+
+		fmt.Printf(color.Green("Success! %s secret was deleted.", "bold", 1), to_get_name)
+
+	case "N", "n", "no", "No", "NO":
+		fmt.Printf(color.White("Ok! %s secret was not deleted.", "bold", 1), to_get_name)
+		return 
+	default:
+		fmt.Printf(color.White("%s secret was not deleted.", "bold", 1), to_get_name)
+		return 
+	}
+
 }
